# openapi.yaml
openapi: 3.0.3
info:
  title: TINTAS Backend API
  version: 1.0.0
  description: |
    Documentacion OpenAPI generada desde las rutas y controladores del backend.
    Se puede pegar directamente en https://editor.swagger.io/.

    Autenticacion: varias rutas usan JWT por cookie `token` o header
    `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:3000
    description: Local
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/user:
    get:
      tags: [Users]
      summary: Listar usuarios
      description: Requiere JWT y rol SUPERADMIN.
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: the users
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/book:
    get:
      tags: [Books]
      summary: Listar libros
      description: Requiere JWT. Si el usuario tiene nivel, devuelve libros habilitados para ese nivel.
      responses:
        "200":
          description: Lista de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/progress:
    get:
      tags: [Progress]
      summary: Obtener progresos del usuario autenticado
      description: El controlador usa `req.user.id`; si no existe, intenta `idUser` desde body.
      responses:
        "200":
          description: Progresos del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookProgressExpanded"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Progress]
      summary: Actualizar progreso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdateRequest"
      responses:
        "200":
          description: Progreso actualizado, o eliminado si la actualizacion no devuelve resultado
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Progress updated successfully.
                  result:
                    $ref: "#/components/schemas/BookProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Progress]
      summary: Eliminar progreso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
      responses:
        "200":
          description: Progreso eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  responses:
...
    Unauthorized:
      description: No autenticado o token invalido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    Forbidden:
      description: Sin permisos suficientes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Role:
      type: string
      enum: [USER, ADMIN, SUPERADMIN]
    Level:
      type: string
      enum: [INICIAL, SECUNDARIO, JOVEN_ADULTO, ADULTO_MAYOR]
    ProgressUnit:
      type: string
      enum: [page, pages, second, seconds]
      description: En base de datos se guarda en mayusculas.
    ProgressStatus:
      type: string
      enum: [reading, finished, pending, abandoned]
      description: En base de datos solo existen READING, FINISHED y PENDING.
    MessageResponse:
      type: object
      properties:
        msg:
          type: string
          example: operation successful
        message:
          type: string
          example: operation successful
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        msg:
          type: string
        message:
          type: string
        error:
          nullable: true
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        lastName:
          type: string
        userName:
          type: string
        birthDate:
          type: string
          format: date-time
        email:
          type: string
          format: email
        password:
          type: string
          description: Hash de password.
        nivel:
          type: string
          nullable: true
        imgLevel:
          type: string
          nullable: true
        level:
          $ref: "#/components/schemas/Level"
        rol:
          $ref: "#/components/schemas/Role"
        point:
          type: integer
        avatar:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Author:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        biography:
          type: string
        profession:
          type: string
        birthdate:
          type: string
          format: date-time
        birthplace:
          type: string
        nationality:
          type: string
        isActivo:
          type: boolean
        photoIdImage:
          type: string
        photoUrl:
          type: string
          format: uri
...
    ContentBook:
      type: object
      properties:
        idContentBook:
          type: string
        url_secura:
          type: string
          format: uri
    BookCover:
      type: object
      properties:
        idBookCoverImage:
          type: string
        url_secura:
          type: string
          format: uri
    Book:
      type: object
      properties:
        id:
          type: string
        _id:
          nullable: true
        __v:
          type: integer
          example: 0
        title:
          type: string
        author:
          type: array
          items:
            $ref: "#/components/schemas/Author"
        summary:
          type: string
        subgenre:
          type: array
          items:
            type: string
        language:
          type: string
        available:
          type: boolean
        yearBook:
          type: string
        synopsis:
          type: string
        theme:
          type: array
          items:
            type: string
        genre:
          type: string
        level:
          $ref: "#/components/schemas/Level"
        format:
          type: string
        totalPages:
          type: integer
          nullable: true
        duration:
          type: integer
          nullable: true
        anthology:
          type: boolean
        fileExtension:
          type: string
        contentBook:
          $ref: "#/components/schemas/ContentBook"
        bookCoverImage:
          $ref: "#/components/schemas/BookCover"
    BookPrisma:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          properties:
            authors:
              type: array
              items:
                $ref: "#/components/schemas/Author"
            contentBookId:
              type: string
            contentBookUrl:
              type: string
            coverImageId:
              type: string
            coverImageUrl:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    ProgressCreateRequest:
      type: object
      required: [idBook, unit, position, percent, total]
      properties:
        idBook:
          type: string
        unit:
          $ref: "#/components/schemas/ProgressUnit"
        position:
          type: integer
          example: 15
        percent:
          type: number
          format: float
          example: 23.5
        total:
          type: integer
          example: 120
        status:
          $ref: "#/components/schemas/ProgressStatus"
        startDate:
          type: string
          format: date-time
        finishDate:
          type: string
          format: date-time
          nullable: true
    ProgressUpdateRequest:
      allOf:
        - type: object
          required: [id]
          properties:
            id:
              type: string
        - $ref: "#/components/schemas/ProgressCreateRequest"
    BookProgress:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bookId:
          type: string
        unit:
          type: string
          enum: [PAGE, PAGES, SECOND, SECONDS]
        position:
          type: integer
        percent:
          type: number
          format: float
        total:
          type: integer
        status:
          type: string
          enum: [READING, FINISHED, PENDING]
        startDate:
          type: string
          format: date-time
        finishDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookProgressExpanded:
      allOf:
        - $ref: "#/components/schemas/BookProgress"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            book:
              $ref: "#/components/schemas/BookPrisma"